+++ packages/select-text/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.10.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text@3.10.0...@pie-element/select-text@3.10.1) (2019-04-09)
+
+**Note:** Version bump only for package @pie-element/select-text
+
+
+
+
+
+# [3.10.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text@3.9.7...@pie-element/select-text@3.10.0) (2019-04-09)
+
+
+### Features
+
+* partial scoring - use @pie-lib/controller-utils ([dac819f](https://github.com/pie-framework/pie-elements/commit/dac819f))
+
+
+
+
+
+## [3.9.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text@3.9.6...@pie-element/select-text@3.9.7) (2019-04-08)
+
+
+### Bug Fixes
+
+* **select-text:** Fixed console warning. ([5957c41](https://github.com/pie-framework/pie-elements/commit/5957c41))
+
+
+
+
+
 ## [3.9.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text@3.9.5...@pie-element/select-text@3.9.6) (2019-04-08)
 
 **Note:** Version bump only for package @pie-element/select-text