+++ packages/select-text/configure/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.5.7...@pie-element/select-text-configure@2.5.8) (2019-04-09)
+
+**Note:** Version bump only for package @pie-element/select-text-configure
+
+
+
+
+
+## [2.5.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.5.6...@pie-element/select-text-configure@2.5.7) (2019-04-08)
+
+
+### Bug Fixes
+
+* **select-text:** Fixed console warning. ([5957c41](https://github.com/pie-framework/pie-elements/commit/5957c41))
+
+
+
+
+
 ## [2.5.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.5.5...@pie-element/select-text-configure@2.5.6) (2019-04-08)
 
 **Note:** Version bump only for package @pie-element/select-text-configure